+++ src/leiningen/new/re_frame/src/cljs/events.cljs
@@ -1,6 +1,7 @@
 (ns {{ns-name}}.events
   (:require
-   [re-frame.core :as re-frame]
+   [re-frame.core :as re-frame]{{#re-pressed?}}
+   [re-pressed.core :as rp]{{/re-pressed?}}
    [{{ns-name}}.db :as db]{{#10x?}}
    [day8.re-frame.tracing :refer-macros [fn-traced]]{{/10x?}}
    ))
@@ -11,10 +12,26 @@
    db/default-db))
 {{#routes?}}
 
-(re-frame/reg-event-db
+(re-frame/reg-event-fx
+  ::navigate
+  ({{^10x?}}fn{{/10x?}}{{#10x?}}fn-traced{{/10x?}} [_ [_ handler]]
+   {:navigate handler}))
+
+(re-frame/reg-event-fx
  ::set-active-panel
- ({{^10x?}}fn{{/10x?}}{{#10x?}}fn-traced{{/10x?}} [db [_ active-panel]]
-   (assoc db :active-panel active-panel)))
+ ({{^10x?}}fn{{/10x?}}{{#10x?}}fn-traced{{/10x?}} [{:keys [db]} [_ active-panel]]
+   {:db (assoc db :active-panel active-panel){{#re-pressed?}}
+    :dispatch [::rp/set-keydown-rules
+               {:event-keys [[[::set-re-pressed-example "Hello, world!"]
+                              [{:keyCode 72} ;; h
+                               {:keyCode 69} ;; e
+                               {:keyCode 76} ;; l
+                               {:keyCode 76} ;; l
+                               {:keyCode 79} ;; o
+                               ]]]
+                :clear-keys
+                [[{:keyCode 27} ;; escape
+                  ]]}]{{/re-pressed?}}}))
 {{/routes?}}
 {{#re-pressed?}}