+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Changed
 
+- Change to bidi and pushy for `+routes`
 - Upgrade karma 6.1.1
 - Upgrade shadow-cljs 2.11.21
 - Upgrade re-frame-10x to 1.0.1